Run TDAMapper on a matrix of lab values

Run TDAMapper on a matrix of lab values

map_tda(lab_values_mat, ...)

Arguments

lab_values_mat
Matrix of (imputed) lab values.

...
Arguments to pass to TDAMapper()

Value

TDAMapper object which is a list of items named adjacency (adjacency matrix for the edges), num_vertices (integer number of vertices), level_of_vertex (vector with level_of_vertex[i] = index of the level set for vertex i), points_in_vertex (list with points_in_vertex[[i]] = vector of indices of points in vertex i), points_in_level (list with points_in_level[[i]] = vector of indices of points in level set i, and vertices_in_level (list with vertices_in_level[[i]] = vector of indices of vertices in level set i.

Examples

my_tda <- map_tda(scaled_lab_mat)
str(my_tda, max.lev = 1)

#> List of 6
#>  $ adjacency        : num [1:26, 1:26] 0 1 0 0 0 1 0 0 0 0 ...
#>  $ num_vertices     : num 26
#>  $ level_of_vertex  : int [1:26] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ points_in_vertex :List of 26
#>  $ points_in_level  :List of 25
#>  $ vertices_in_level:List of 25
#>  - attr(*, "class")= chr "TDAmapper"